Update primary data insertion function
populate_db now accepts a source of data rather than generating one
internally.  This allows me to feed it with mock data in order to test
its functionality in a controlled fashion.

Since this represents a very different way of doing things, I had to
update the CLI command as well.