Bugfixes: get flower to work and add network
Flower is now running properly
The network issue is a dev-only problem: I noticed that docker was
creating a default network in addition to toolhunt-net and that, I
presume, is what the celery, redis and flower containers were connecting
to.  I'm not sure whether that might be a problem, but to make sure that
it never becomes one, I explicitly placed the new containers on toolhunt-net.