Bugfix: Format data values for multi-choice fields
This addresses https://phabricator.wikimedia.org/T331262

As I wasn't having any luck creating a custom schema that would
correctly handle both strings and lists, I opted to accept a string
and convert it into a list.

Unfortunately, Flask did not do a good job of converting the resulting
data to a valid JSON, so I had to do it manually.