Commit complete rewrite of db population functions
This was 99% Slavina's work; I only tweaked some of it.

Notably, upserting tasks did not work, possibly because the insert was
not encountering any unique values and therefore did not recognize the
existence of a duplicate task.  I have fallen back on the standby of
checking for the existence of an existing task before inserting.